libxl: define a per-domain logger.
libxl doesn't provide a way to write one log for each domain. Thus
we need to demux the messages. If our logger doesn't know to which
domain to attribute a message, then it will write it to the default
log file.

So far libxl doesn't write the domain id to the log at all. This will
change, but with the default log file no log message will be lost.